canvas-view: improve usability of sort options
The sort options shown on the toolbar menu aren't very intuitive. For
example, 'last modified, reverse order' can be interpreted a number of
ways. They also require the user to set both the sort order and whether
the sort should be reversed. These issues have been addressed by
replacing the sort order items that are often sorted in either ascending
and descending order (e.g. name) with options that are better to use
(e.g. A-Z and Z-A). Options that are generally only sorted one way
(e.g. type) have been changed to sort in ascending/descending as
appropriate for the option.

This is part of the toolbar menu redesign to improve the usability of
the menus.

https://bugzilla.gnome.org/show_bug.cgi?id=764632