toolbar: move undo/redo toolbar menu code into toolbar
Due to the toolbar menu reorganisation work, the code to create and
manage the undo/redo items on the menu ended up in files-view.c. This
isn't the correct place as they don't have much to do with the files
view. Some refactoring was needed before the code for these items could
be moved back into the toolbar, which has now been done in a previous
commit.

This commit moves the undo/redo creation and management code back into
the toolbar (and partially into nautilus-window.c).

https://bugzilla.gnome.org/show_bug.cgi?id=764632